fix(gui): board filter count reflects collapsed cards; merge board tests into taskFilters
Review round-1 fixes for the W8 ordering/collapse change: the filter bar
and board header now report the visible card count (boardTasks) instead
of the pre-collapse filter result, so collapsed cold terminal tasks no
longer inflate the displayed count while the independent cold-count
switch stays visible. The board/swimlane component tests move into the
existing taskFilters.vitest.ts (manifest-kept runner coverage; the
extra test file and gui-test-manifest.mjs entry are reverted), and a
same-markup assertion proves the count matches rendered cards in both
switch states. viewHistoryStorage drops the missing-field tolerance for
expandColdTerminal: stored histories without the boolean are invalid
locations and fall back to the clean initial stack.
